Adicioando tela de carrinho de compras.

+++ FelipeShopping/FelipeShopping.Web/Models/CarrinhoHeaderViewModel.cs
@@ -7,6 +7,6 @@
 
         public string CodigoCupom { get; set; }
 
-        public double ValorCompra { get; set; }
+        public decimal ValorCompra { get; set; }
     }
 }
